+++ components/home/Tools/index.js
@@ -1,13 +1,30 @@
 import React from 'react';
+import ScaletorLogotype from '../../scaletor/Logotype';
+
+const items = [
+  {
+    id: 'item-scaletor',
+    link: {
+      label: 'Scaletor',
+      url: '/scaletor',
+    },
+    image: <ScaletorLogotype className="tools__items-image" />,
+  },
+];
 
 const Tools = () => (
   <section className="tools">
-    <a
-      className="tools__items"
-      href="/scaletor"
-    >
-      <img className="header__logo" src="/images/scaletor/general/logotype.svg" alt="" />
-    </a>
+    {
+      items.map((item) => (
+        <a
+          id={item.id}
+          className="tools__items"
+          href={item.link.url}
+        >
+          {item.image}
+        </a>
+      ))
+    }
   </section>
 );