+++ components/scaletor/Example/index.js
@@ -1,9 +1,0 @@
-import React from 'react';
-
-const Example = () => (
-  <div className="c-example">
-    <h1>Example</h1>
-  </div>
-);
-
-export default Example;