+++ CHANGELOG.md
@@ -2,9 +2,12 @@
 All notable changes to this project will be documented in this file.
 
 ## [Unreleased]
-### Changes
 
-## [Unreleased]
+## [1.1.0]
+### Changes
+- Update images implementation [#4](https://github.com/fdrandolfi/composition-tools-frontend/pull/4)
+- Add commons docs, favicons & Update Scaletor [#3](https://github.com/fdrandolfi/composition-tools-frontend/pull/3)
 
 ## [1.0.0]
-- Add Main Page & Migrate Scaletor Project on New Page [#1](https://github.com/fdrandolfi/next-boilerplate/pull/1)
+### Changes
+- Add Main Page & Migrate Scaletor Project on New Page [#1](https://github.com/fdrandolfi/composition-tools-frontend/pull/1)